requirements.txt
flask
pyqrcode
Pillow
pypng